Readme.me
Here is a chatroom written with C# Lang Run the followings in order:

socketporgraming.exe
Client_Side.exe
 

Now click connect butten on server and on client enter username and click on connect. 
if you want to chat with every one do not enter contact and if you want to have private chat type username in contact textbox.